Refresh ExifTool output to interpret updated namespaces
The main reason for this patch is the change of the RDF namespace
prefixes from `ns.exiftool.ca` to `ns.exiftool.org` with version 12.26.

This patch regenerates the ExifTool output on the govdocs1 sample.
One manual edit-pattern was applied: File system timestamps were
preserved from the prior Git-tracked state.  Otherwise, the RDF-XML
results are as generated.

A follow-on patch will regenerate Make managed files.

References:
* https://exiftool.org/ancient_history.html#v12.26

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>