Add `pdfCreationDate` and `pdfModDate` to `PDFFile` mapping
The patch adds designations for MIME types, `application/pdf` and
`image/jpeg`, with UCO classes, `PDFFile` and `RasterPicture`.

The patch maps two concepts that already exist in UCO,
`pdfCreationDate` and `pdfModDate`.

The patch adds IRI reference to the `ExifTool` PDF namespace and binds
to graph model.

The patch hardcodes `self._oo_slug` to be 'File-'. Representation as a
file object encompasses accepted format types of `ExifTool`.

The patch provides a test file, `test_000015.py`, which validates the
presence of the `PDFFile` class.

Reference
* https://github.com/ucoProject/UCO/issues/421

Signed-off-by: Sheldon Douglas <sheldon.douglas@nist.gov>